requirements.txt
absl-py==1.0.0
aiofiles==0.8.0
aiohttp==3.8.1
aiosignal==1.2.0
alabaster==0.7.12
albumentations==1.3.0
anaconda-client==1.7.2
anaconda-navigator==1.9.12
anaconda-project==0.8.3
analytics-python==1.4.0
aniso8601==9.0.1
antlr4-python3-runtime==4.8
anyio==3.5.0
appdirs==1.4.4
APScheduler==3.6.3
argh==0.26.2
asgiref==3.5.0
asn1crypto==1.3.0
astroid==2.3.3
astropy==4.0
astunparse==1.6.3
async-timeout==4.0.2
asynctest==0.13.0
atomicwrites==1.3.0
attrs==19.3.0
audioread==2.1.9
autopep8==1.4.4
Babel==2.8.0
backcall==0.1.0
backoff==1.10.0
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
backports.zoneinfo==0.2.1
bcrypt==3.2.0
beautifulsoup4==4.8.2
bitarray==1.2.1
bkcharts==0.2
black==21.4b2
bleach==3.1.0
blobfile==1.2.8
bokeh==1.4.0
boto==2.49.0
Bottleneck==1.3.2
cachetools==4.2.2
certifi==2022.12.7
cffi==1.14.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.0
cloudpickle==1.3.0
clyent==1.2.2
colorama==0.4.3
colorlog==6.7.0
conda==4.8.2
conda-build==3.18.11
conda-package-handling==1.6.0
conda-verify==3.4.2
contextlib2==0.6.0.post1
cryptography==2.8
cycler==0.10.0
Cython==0.29.15
cytoolz==0.10.1
dask==2.11.0
decorator==4.4.1
deepspeed==0.4.0
defusedxml==0.6.0
detectron2==0.6+cu102
diff-match-patch==20181111
distributed==2.11.0
dlib==19.23.1
dnspython==2.2.1
docutils==0.16
easydict==1.10
einops==0.4.1
entrypoints==0.3
et-xmlfile==1.0.1
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.75.0
fastcache==1.1.0
ffmpeg==1.4
ffmpy==0.3.0
ffprobe==0.5
filelock==3.0.12
flake8==3.7.9
Flask==1.1.1
Flask-Cors==3.0.10
Flask-JWT-Simple==0.0.3
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.9
flatbuffers==2.0
fonttools==4.38.0
frozenlist==1.3.0
fsspec==0.6.2
future==0.18.2
fvcore==0.1.5.post20220305
gast==0.5.3
gevent==1.4.0
gitdb==4.0.10
GitPython==3.1.29
glob2==0.7
gmpy2==2.0.8
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gradio==2.9.0
greenlet==0.4.15
grpcio==1.44.0
h11==0.13.0
h5py==2.10.0
HeapDict==1.0.1
hnswlib==0.6.2
html5lib==1.0.1
huggingface-hub==0.4.0
hydra-core==1.1.1
hypothesis==5.5.4
idna==2.8
imageio==2.6.1
imageio-ffmpeg==0.4.5
imagesize==1.2.0
importlib-metadata==4.11.3
importlib-resources==5.6.0
imutils==0.5.4
insightface==0.4.2
intervaltree==3.0.2
iopath==0.1.9
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.14.1
jeepney==0.4.2
Jinja2==2.11.1
joblib==0.14.1
json5==0.9.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
jupyterlab==1.2.6
jupyterlab-server==1.0.6
keras==2.8.0
Keras-Preprocessing==1.1.2
keyring==21.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
libarchive-c==2.8
libclang==13.0.0
librosa==0.9.1
lief==0.9.0
lightgbm==3.3.2
linkify-it-py==1.0.3
llvmlite==0.31.0
locket==0.2.0
lxml==4.5.0
Markdown==3.3.6
markdown-it-py==2.0.1
MarkupSafe==1.1.1
marshmallow==3.18.0
matplotlib==3.5.3
mccabe==0.6.1
mdit-py-plugins==0.3.0
mdurl==0.1.0
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mock==4.0.1
mongoengine==0.24.2
monotonic==1.6
more-itertools==8.2.0
moviepy==1.0.3
mpmath==1.1.0
msgpack==0.6.1
multidict==6.0.2
multipledispatch==0.6.0
mypy==0.941
mypy-extensions==0.4.3
navigator-updater==0.2.1
nbconvert==5.6.1
nbformat==5.0.4
nest-asyncio==1.5.4
networkx==2.4
newsapi==0.1.1
ninja==1.10.2.3
nltk==3.4.5
nose==1.3.7
notebook==6.0.3
numba==0.48.0
numexpr==2.7.1
numpy==1.21.5
numpydoc==0.9.2
oauthlib==3.2.0
olefile==0.46
omegaconf==2.1.1
onnx==1.13.0
onnxruntime-gpu==1.9.0
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
openpyxl==3.0.3
opt-einsum==3.3.0
orjson==3.6.7
packaging==21.3
pandas==1.0.1
pandocfilters==1.4.2
paramiko==2.10.3
parso==0.5.2
partd==1.1.0
path==13.1.0
pathlib2==2.3.5
pathspec==0.9.0
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pkginfo==1.5.0.1
pluggy==0.13.1
ply==3.11
pooch==1.6.0
portalocker==2.4.0
pose-format==0.1.2
prettytable==3.5.0
proglog==0.1.9
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.20.3
psutil==5.6.7
ptyprocess==0.6.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycryptodome==3.14.1
pycryptodomex==3.14.1
pycurl==7.43.0.5
pydantic==1.9.0
pydocstyle==4.0.1
pydot==1.4.2
pydub==0.25.1
pyflakes==2.1.1
Pygments==2.5.2
PyJWT==2.5.0
pylint==2.4.4
pymongo==4.1.0
PyNaCl==1.5.0
pyodbc===4.0.0-unsupported
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyrsistent==0.15.7
PySocks==1.7.1
pytest==5.3.5
pytest-arraydiff==0.3
pytest-astropy==0.8.0
pytest-astropy-header==0.1.2
pytest-doctestplus==0.5.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
python-dateutil==2.8.1
python-dotenv==0.21.0
python-jsonrpc-server==0.3.4
python-language-server==0.31.7
python-multipart==0.0.5
python-telegram-bot==13.15
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.1.1
pyxdg==0.26
PyYAML==5.3
pyzmq==18.1.1
QDarkStyle==2.8
QtAwesome==0.6.1
qtconsole==4.6.0
QtPy==1.9.0
qudida==0.0.4
regex==2022.3.15
requests==2.22.0
requests-oauthlib==1.3.1
resampy==0.2.2
rope==0.16.0
rsa==4.8
Rtree==0.9.3
ruamel-yaml==0.15.87
sacremoses==0.0.49
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
SecretStorage==3.1.2
Send2Trash==1.5.0
Shapely==1.8.1.post1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
SoundFile==0.10.3.post1
soupsieve==1.9.5
Sphinx==2.4.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.2.0
spyder==4.0.1
spyder-kernels==1.8.1
SQLAlchemy==1.3.13
starlette==0.17.1
statsmodels==0.11.0
sympy==1.5.1
tables==3.6.1
tabulate==0.8.9
tblib==1.6.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tf-estimator-nightly==2.8.0.dev2021122109
timm==0.5.4
tokenizers==0.11.6
toml==0.10.2
tomli==2.0.1
toolz==0.10.0
torch==1.10.1+cu111
torchvision==0.11.2+cu111
tornado==6.1
tqdm==4.42.1
traitlets==4.3.3
transformers==4.17.0
triton==1.1.1
typed-ast==1.4.3
typing-extensions==4.1.1
tzdata==2022.7
tzlocal==4.2
uc-micro-py==1.0.1
ujson==1.35
unicodecsv==0.14.1
urllib3==1.25.8
uvicorn==0.17.6
vidgear==0.2.6
watchdog==0.10.2
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wrapt==1.11.2
wurlitzer==2.0.0
xlrd==1.2.0
XlsxWriter==1.2.7
xlwt==1.3.0
xmltodict==0.12.0
yacs==0.1.8
yapf==0.28.0
yarl==1.7.2
zict==1.0.0
zipp==3.7.0
